split data.ts
