+++ src/_data/header.ts
@@ -1,0 +1,6 @@
+import { HeaderDataType } from "../domain";
+
+export const headerData: HeaderDataType = {
+  name: "Sora Ichigo",
+  description: "software engineer",
+};
