+++ src/_data/index.ts
@@ -1,0 +1,4 @@
+export * from "./header";
+export * from "./about";
+export * from "./resume";
+export * from "./portfolio";
