+++ src/common/utils/mainPages.ts
@@ -1,6 +1,6 @@
 import React from "react";
 import { About } from "../../about/components/About";
-import { aboutData, portfolioData, resumeData } from "../../data";
+import { aboutData, portfolioData, resumeData } from "../../_data";
 import { Data } from "../../domain";
 import { Portfolio } from "../../portfolio/components/Portfolio";
 import { Resume } from "../../resume/components/Resume";
